@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -7,34 +8,21 @@ using Microsoft.Extensions.Logging;
 
 namespace HappyCode.NetCoreBoilerplate.Api.Infrastructure.Filters
 {
-    public class HttpGlobalExceptionFilter : IExceptionFilter
+    [ExcludeFromCodeCoverage]
+    public class HttpGlobalExceptionFilter(IWebHostEnvironment env, ILogger<HttpGlobalExceptionFilter> logger) : IExceptionFilter
     {
-        private readonly IWebHostEnvironment _env;
-        private readonly ILogger<HttpGlobalExceptionFilter> _logger;
-
-        public HttpGlobalExceptionFilter(IWebHostEnvironment env, ILogger<HttpGlobalExceptionFilter> logger)
-        {
-            _env = env;
-            _logger = logger;
-        }
-
         public void OnException(ExceptionContext context)
         {
-            _logger.LogError(new EventId(context.Exception.HResult),
-                context.Exception,
-                context.Exception.Message);
-
-            var jsonErrorResponse = new ErrorResponse
+            if (env.IsDevelopment())
             {
-                Messages = new[] { "An internal error has occurred" }
-            };
-
-            if (_env.IsDevelopment())
-            {
-                jsonErrorResponse.Exception = context.Exception.ToString();
+                throw context.Exception;
             }
 
-            context.Result = new ObjectResult(jsonErrorResponse) { StatusCode = StatusCodes.Status500InternalServerError };
+            logger.LogError(context.Exception, "An API internal error has occurred");
+
+            var errorResponse = new HttpValidationProblemDetails { Title = "An internal error has occurred" };
+            context.Result = new ObjectResult(errorResponse) { StatusCode = StatusCodes.Status500InternalServerError };
+
             context.ExceptionHandled = true;
         }
     }
