@@ -1,7 +1,6 @@
 using HappyCode.NetCoreBoilerplate.Api;
 using HappyCode.NetCoreBoilerplate.Api.Infrastructure.Configurations;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Builder;
 using Serilog;
 
 Log.Logger = SerilogConfigurator.CreateLogger();
@@ -11,8 +10,12 @@ try
     BannerConfigurator.Print(!Console.IsOutputRedirected);
 
     Log.Logger.Information("Starting up...");
-    using var webHost = CreateWebHostBuilder(args).Build();
-    await webHost.RunAsync();
+    var builder = WebApplication.CreateBuilder();
+    var startup = new Startup(builder.Configuration);
+    startup.ConfigureServices(builder.Services);
+    using var app = builder.Build();
+    startup.Configure(app, builder.Environment);
+    await app.RunAsync();
     Log.Logger.Debug("I'm done, see ya later!");
 }
 catch (Exception ex)
@@ -24,11 +27,3 @@ finally
 {
     Log.CloseAndFlush();
 }
-
-static IHostBuilder CreateWebHostBuilder(string[] args) =>
-    Host.CreateDefaultBuilder(args)
-        .UseSerilog()
-        .ConfigureWebHostDefaults(webBuilder =>
-        {
-            webBuilder.UseStartup<Startup>();
-        });
