@@ -1,6 +1,5 @@
 using HappyCode.NetCoreBoilerplate.Api.BackgroundServices;
-using HappyCode.NetCoreBoilerplate.Api.Infrastructure.Filters;
-using HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure.DataFeeders;
+using HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure.DataSeeders;
 using HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure.Fakes;
 using HappyCode.NetCoreBoilerplate.Core;
 using HappyCode.NetCoreBoilerplate.Core.Registrations;
@@ -9,22 +8,14 @@ using Microsoft.FeatureManagement;
 
 namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
 {
-    internal class TestStartup : Startup
+    internal class TestStartup(IConfiguration configuration)
+        : Startup(configuration)
     {
-        public TestStartup(IConfiguration configuration)
-            : base(configuration)
-        {
-
-        }
-
         public override void ConfigureServices(IServiceCollection services)
         {
             services
                 .AddHttpContextAccessor()
-                .AddMvcCore(options =>
-                {
-                    options.Filters.Add<ValidateModelStateFilter>();
-                })
+                .AddMvcCore()
                 .AddDataAnnotations();
 
             services.AddCoreComponents();
@@ -34,21 +25,24 @@ namespace HappyCode.NetCoreBoilerplate.Api.IntegrationTests.Infrastructure
 
             services.AddDbContext<EmployeesContext>(options =>
             {
-                options.UseInMemoryDatabase("employees");
+                options.UseInMemoryDatabase("employees")
+                    .UseSeeding(EmployeesDataSeeder.Seed)
+                    .UseAsyncSeeding(EmployeesDataSeeder.SeedAsync);
             });
             services.AddDbContext<CarsContext>(options =>
             {
-                options.UseInMemoryDatabase("cars");
+                options.UseInMemoryDatabase("cars")
+                    .UseSeeding(CarsDataSeeder.Seed)
+                    .UseAsyncSeeding(CarsDataSeeder.SeedAsync);
             });
         }
 
         public override void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            var employeesContext = app.ApplicationServices.GetService<EmployeesContext>();
-            EmployeesContextDataFeeder.Feed(employeesContext);
-
-            var carsContext = app.ApplicationServices.GetService<CarsContext>();
-            CarsContextDataFeeder.Feed(carsContext);
+            using var employees = app.ApplicationServices.GetRequiredService<EmployeesContext>();
+            employees.Database.EnsureCreated();
+            using var cars = app.ApplicationServices.GetRequiredService<CarsContext>();
+            cars.Database.EnsureCreated();
 
             app.UseRouting();
             app.UseEndpoints(endpoints =>
